Try to use 'interruptProcessOf' to interrupt the running process.
This does not really work as the whole process group is terminated. Moreover,
we cannot get a handle of the actual GHCi process, as it is started trough a
shell. So, it might not be really easy to send a SIGINT directly to it.

Another option is for the tests to listen on a socket for a kill signal, and
have GHCid submit it there.